requirements.txt
Flask>=2.0.0
pywebview>=3.0.0
